refactor: type 파일 분리와 쿼리 타입 통합 및 관련 import 정리
